services/auth_service/requirements.txt
fastapi
uvicorn
python-jose
passlib
psycopg2-binary
pydantic
